iot/requirements.txt
boto3==1.36.16
keras==3.3.0
paho-mqtt==2.1.0
python-dotenv==1.0.1
tensorflow==2.17.0
opencv-python==4.10.0.84